rename get methods with fetch
extract data fetching functions

extract commit processing functions

extract print functions

fix name

export commit types from commits.py
