+++ cache/codes/LEGITEXT000006070208.json
@@ -1,5 +1,4 @@
 {
-    "executionTime": 6,
     "dereferenced": false,
     "id": "LEGITEXT000006070208_26-05-2023",
     "idConteneur": null,
