+++ cache/codes/LEGITEXT000006070239.json
@@ -1,5 +1,4 @@
 {
-    "executionTime": 7,
     "dereferenced": false,
     "id": "LEGITEXT000006070239_21-07-2023",
     "idConteneur": null,
