+++ cache/codes/LEGITEXT000006072051.json
@@ -1,5 +1,4 @@
 {
-    "executionTime": 3,
     "dereferenced": false,
     "id": "LEGITEXT000006072051_25-09-2020",
     "idConteneur": null,
