+++ cache/codes/LEGITEXT000006074073.json
@@ -1,5 +1,4 @@
 {
-    "executionTime": 6,
     "dereferenced": false,
     "id": "LEGITEXT000006074073_07-06-2023",
     "idConteneur": null,
