+++ cache/codes/LEGITEXT000006074228.json
@@ -1,5 +1,4 @@
 {
-    "executionTime": 5,
     "dereferenced": false,
     "id": "LEGITEXT000006074228_01-09-2023",
     "idConteneur": null,
