+++ cache/codes/LEGITEXT000044595989.json
@@ -1,5 +1,4 @@
 {
-    "executionTime": 5,
     "dereferenced": false,
     "id": "LEGITEXT000044595989_12-07-2023",
     "idConteneur": null,
